feature: convertido codigo jdbc para jpa do sistema de emprestimos

+++ src/main/java/br/com/nyk/sgb/DTO/EmprestimoCliente.java
@@ -1,5 +1,5 @@
 package br.com.nyk.sgb.DTO;
 
 public record EmprestimoCliente(String nome,
-                                Integer quantidade) {
+                                Long quantidade) {
 }
