+++ src/main/java/br/com/nyk/sgb/DTO/EmprestimoTopCinco.java
@@ -1,5 +1,5 @@
 package br.com.nyk.sgb.DTO;
 
 public record EmprestimoTopCinco(String titulo,
-                                 Integer contagem) {
+                                 Long contagem) {
 }
